Delete with approve

+++ source/webapp/views.py
@@ -28,5 +28,8 @@ def view(request, *args, pk):
 
 def delete(request, *args, pk):
     task = get_object_or_404(Task, pk=pk)
-    task.delete()
-    return redirect('index')
+    if request.method == "GET":
+        return render(request, 'task_delete.html', {'task': task})
+    elif request.method == "POST":
+        task.delete()
+        return redirect('index')
